refactor: modularize event-handler transform functions
