+++ packages/core/default-mcfly-config.js
@@ -1,0 +1,3 @@
+export default {
+  components: 'js',
+}
