+++ packages/core/hooks.js
@@ -1,0 +1,6 @@
+export const hooks = {
+  pageRendered: 'mcfly:page:rendered',
+  serverScriptsEvaluated: 'mcfly:scripts:evaluated',
+  fragmentsInjected: 'mcfly:fragments:injected',
+  customElementsInjected: 'mcfly:elements:injected',
+}
