+++ packages/core/package.json
@@ -25,12 +25,13 @@
     "consola": "^3.3.3",
     "esbuild": "^0.24.2",
     "esprima": "^4.0.1",
-    "h3": "^1.8.2",
-    "nitropack": "~2.10.4",
+    "h3": "^1.13.0",
+    "nitropack": "latest",
     "pathe": "^1.1.2",
-    "ultrahtml": "^1.5.2"
+    "ultrahtml": "^1.5.3"
   },
   "devDependencies": {
+    "hookable": "^5.5.3",
     "unstorage": "^1.14.4"
   }
 }
